Extracted startup authentication registration.
